refactor: split resources module
