+++ src/deezer/resources/__init__.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+from .album import Album
+from .artist import Artist
+from .chart import Chart
+from .editorial import Editorial
+from .episode import Episode
+from .genre import Genre
+from .playlist import Playlist
+from .podcast import Podcast
+from .radio import Radio
+from .resource import Resource
+from .track import Track
+from .user import User
+
+__all__ = [
+    "Album",
+    "Artist",
+    "Chart",
+    "Editorial",
+    "Episode",
+    "Genre",
+    "Playlist",
+    "Podcast",
+    "Radio",
+    "Resource",
+    "Track",
+    "User",
+]
